+++ examples/hello/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name    = "new_test"
+name    = "hello"
 version = "0.1.0"
 edition = "2021"
